make cov happy
