+++ tests/core/test_config.py
@@ -1,10 +1,8 @@
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
-from unittest.mock import ANY
 
-from src.core.config import AppConfig, PostgresConfig
-from src.core.db import get_db_engine
+from src.core.config import AppConfig
 
 if TYPE_CHECKING:
     from _pytest.monkeypatch import MonkeyPatch
@@ -19,25 +17,3 @@ def test_app_config_multiple_allowed_hosts(monkeypatch: MonkeyPatch) -> None:
 
     # then
     assert config.allowed_hosts == ("example.com", "other.com")
-
-
-async def test_engine_connect_args() -> None:
-    # given
-    config = PostgresConfig(user="test-user", password="test-password", host="test-host", db="test-db")
-
-    # when
-    engine = get_db_engine(config)
-
-    # then
-    args, kwargs = engine.dialect.create_connect_args(engine.url)
-
-    assert args == []
-    assert kwargs == {
-        "connect_timeout": "10",
-        "context": ANY,
-        "dbname": "test-db",
-        "host": "test-host",
-        "password": "test-password",
-        "port": 5432,
-        "user": "test-user",
-    }
